split mpiedm from shared
